add: enemy types

+++ core/src/main/java/com/IONA/TowerDefense/model/Movable.java
@@ -5,7 +5,7 @@ import java.awt.*;
 public interface Movable {
     Point getPosition();
 
-    float getSpeed();
+    int getSpeed();
 
     void setPosition(Point position);
